SCUMM: Proper scaling fix for Backyard Basketball
This is the definite closure for ticket #2729: "I also guess that we
have the same potential problem in the AKOS renderer. At least in
code1_genericDecode()."

Commit 8d23d0799dcacdbc68a13094174f9c5dadb6b3ca fixed one specific
behavior (negative values) however I believe it is just a side effect of
not handling the small costume table in AkosRenderer the same way as in
it's done in ClassicCostumeRenderer.

ClassicCostumeRenderer still uses the byte variables and not the mask
but I put the mask value there already for future use.